requirements.txt
asgiref==3.7.2
crispy-tailwind==1.0.0
Django==5.0.1
django-crispy-forms==2.1
django-environ==0.11.2
django-tailwind==3.8.0
gunicorn==21.2.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.4
